+++ console/main.py
@@ -5,12 +5,12 @@ import rclpy
 from PyQt6.QtWidgets import QApplication
 
 from .control_ui import AP1DebugUI
-from .node import AP1SystemInterfaceNode
+from .node import AP1ConsoleNode
 
 # main time bbgl
 def main(args=None):
     rclpy.init(args=args)
-    ros_node = AP1SystemInterfaceNode()
+    ros_node = AP1ConsoleNode()
     app = AP1DebugUI(ros_node, QApplication(sys.argv))
 
     try: